requirements.txt
pytest~=6.2.4
requests~=2.25.1
Pillow~=8.3.1